[layout] Fix self-collapsing end-margins contributing to overflow.
Previously we just took the end margin-strut, and used this for the
scrollable-overflow. This was incorrect.

As the end margin-strut is used for positioning self-collapsing blocks
we ended up "double counting" these margins.

This patch uses the "final" margin-strut, and subtracts the
self-collapsing end margin-strut to correctly determine the block-end
margin which contributes to scrollable-overflow.

Bug: 1136403
Change-Id: I7d0d0fe801948c3ce818719f28f1a7c795a7ec56