[css-flex] Fix min-height: auto for replaced elements
min-height: auto had been using the image's natural height as the
content size suggestion, ignoring a specified definite inline size, but
according to this spec chain, the specified definite inline size should
transfer through the aspect ratio and affect the min-content size.

https://drafts.csswg.org/css-sizing-3/#min-content ->
https://drafts.csswg.org/css-sizing-3/#intrinsic-sizes ->
https://www.w3.org/TR/CSS2/visudet.html#float-replaced-width ->
https://www.w3.org/TR/CSS2/visudet.html#inline-replaced-width ->

"""
if 'width' has a computed value of 'auto', 'height' has some other
computed value, and the element does have an intrinsic ratio; then the
used value of 'width' is:

(used height) * (intrinsic ratio)
"""

This fix also made clear that the regression caused by
https://chromium-review.googlesource.com/c/chromium/src/+/2333530 was
not so harmless -- the issue wasn't a wrong test in align-items-007, but
rather an incorrect block size given to a replaced element with aspect
ratio but no size in the content size suggestion. We had been giving
such element a 150px intrinsic block size; the correct size is the
available inline size passed through the aspect ratio. This is now fixed
in this patch.

The changed case at the end of flex-aspect-ratio-img-column-011.html was
asserting the opposite behavior of
flex-minimum-height-flex-items-007.xht

Change-Id: I05cb4b3ca796b3df02dfa5d8e3790d6164213be9
Fixed: 7